requirements.txt
aiogram>=3.26.0
aiohttp>=3.13.3
Pillow>=12.1.1
pandas>=3.0.1
openpyxl>=3.1.0
python-decouple>=3.8
aiosqlite>=0.20.0
